# API/openapi.yaml
openapi: 3.0.3
info:
  title: UnlikeOtherAuthenticator API
  version: 1.0.0
  description: |
    Centralized OAuth / authentication service. Provides user auth flows,
    token issuance, organisation/team management, and admin operations.

    ## Authentication Schemes

    | Scheme | Header | Format | Used By |
    |--------|--------|--------|---------|
    | Domain Hash Bearer | `Authorization` | `Bearer SHA256(normalizedDomain+clientSecret)` | Domain-scoped server-to-server calls |
    | Admin Bearer | `Authorization` | `Bearer <HS256 JWT>` | Internal admin endpoints |
    | Access Token | `X-UOA-Access-Token` | `Bearer <JWT>` or raw JWT | User-context org endpoints |
    | Config JWT | `X-UOA-Config-JWT` | RS256 JWT | Email send endpoint |

    ## Config URL
    Most endpoints require a `config_url` query parameter pointing to a signed
    RS256 config JWT hosted by the integrator. The server fetches and verifies
    this JWT on every request.
servers:
  - url: https://auth.example.com
    description: Production
security: []
...
components:
  securitySchemes:
    domainHashBearer:
      type: http
      scheme: bearer
      description: |
        `Bearer SHA256(normalizedDomain + clientSecret)` — uses the per-domain client secret, not the shared secret.
    adminBearer:
      type: http
      scheme: bearer
      description: |
        HS256 JWT signed with `ADMIN_ACCESS_TOKEN_SECRET`. Claims must include
        `role: "superuser"` and `domain` matching `ADMIN_AUTH_DOMAIN`.
    accessToken:
      type: apiKey
      in: header
      name: X-UOA-Access-Token
      description: |
        UOA access token (HS256 JWT). May be prefixed with `Bearer ` or sent raw.
    configJwt:
      type: apiKey
      in: header
      name: X-UOA-Config-JWT
      description: RS256 config JWT signed with the domain's private key.
    billingAppKey:
      type: apiKey
      in: header
      name: X-UOA-App-Key
      description: One revocable UOA app key dedicated to the exact calling product/deployment and either entitlement or customer-lifecycle endpoints.
    billingActor:
      type: apiKey
      in: header
      name: X-UOA-Actor
      description: Short-lived RS256 actor assertion bound to the exact product app key and UOA user/org/team.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe endpoint signature verified over the exact raw request body with a separate webhook secret.
  parameters:
    configUrl:
      name: config_url
      in: query
      required: true
      description: HTTPS URL to a signed RS256 config JWT hosted by the integrator.
      schema:
        type: string
        format: uri
    domainQuery:
      name: domain
      in: query
      required: true
      description: Normalised domain (lowercase, trimmed).
      schema:
        type: string
        example: acme.example.com
    limitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
    cursorQuery:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    orgIdPath:
      name: orgId
      in: path
      required: true
      schema:
        type: string
    teamIdPath:
      name: teamId
      in: path
      required: true
      schema:
        type: string
    groupIdPath:
      name: groupId
      in: path
      required: true
      schema:
        type: string
...
    avatarStyleQuery:
      name: style
      in: query
      required: false
      description: |
        Overrides the generated-avatar style for this call. Only affects generated
        avatars — uploaded and proxied provider images are served as stored.
      schema:
        type: string
        enum: [tiles, waves, rings, mono]
    avatarSizeQuery:
      name: size
      in: query
      required: false
      description: |
        Rendered width/height of a generated SVG, clamped to 16-512 (default 128).
        The viewBox is constant; ignored for raster images.
      schema:
        type: integer
        minimum: 16
        maximum: 512
        default: 128
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Generic public error message.
        code:
          type: string
          description: Machine-readable error code (only when DEBUG_ENABLED).
        summary:
          type: string
        details:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
...
    OkTrue:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    PaginatedList:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        next_cursor:
          type: string
          nullable: true
...
    GroupObject:
      type: object
      additionalProperties: true
...
    AvatarUploadResponse:
      type: object
      required: [ok, avatar]
      properties:
        ok:
          type: boolean
          enum: [true]
        avatar:
          type: object
          required: [source, content_type, size_bytes, updated_at]
          properties:
            source:
              type: string
              enum: [uploaded]
            content_type:
              type: string
              enum: [image/png, image/jpeg, image/webp]
              description: The magic-byte sniffed type that was stored.
            size_bytes:
              type: integer
            updated_at:
              type: string
              format: date-time
    AvatarUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: |
            PNG, JPEG or WebP, at most 1 MiB. The stored type is decided by magic-byte
            sniffing, not the supplied mimetype; SVG and other non-raster uploads are
            rejected with the generic error envelope.
...
paths:
...
  /org/organisations/{orgId}/groups:
    parameters:
      - $ref: '#/components/parameters/orgIdPath'
      - $ref: '#/components/parameters/domainQuery'
      - $ref: '#/components/parameters/configUrl'

    get:
      tags: [Organizations]
      summary: List organisation groups
      operationId: listOrgGroups
      description: Requires `org_features.groups_enabled` in config.
      security:
        - domainHashBearer: []
          accessToken: []
      parameters:
        - $ref: '#/components/parameters/limitQuery'
        - $ref: '#/components/parameters/cursorQuery'
      responses:
        '200':
          description: Paginated group list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList'
        '401':
          description: Missing auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Groups feature not enabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /org/organisations/{orgId}/groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/orgIdPath'
      - $ref: '#/components/parameters/groupIdPath'
      - $ref: '#/components/parameters/domainQuery'
      - $ref: '#/components/parameters/configUrl'

    get:
      tags: [Organizations]
      summary: Get group
      operationId: getOrgGroup
      description: Requires `org_features.groups_enabled` in config.
      security:
        - domainHashBearer: []
          accessToken: []
      responses:
        '200':
          description: Group details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupObject'
        '401':
          description: Missing auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found or feature not enabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Internal Admin ────────────────────────────────────────────────────────────
...
  /internal/admin/teams/{teamId}/avatar:
    parameters:
      - $ref: '#/components/parameters/teamIdPath'
    get:
      tags: [Avatars]
      summary: Get any team's avatar image (admin view)
      operationId: adminGetTeamAvatar
      description: |
        Image bytes for any team through the same resolution pipeline as the product-facing
        routes. Requires a bearer credential, so a plain `<img src>` cannot call it — fetch
        the blob and render an object URL.
      security:
        - adminBearer: []
      parameters:
        - $ref: '#/components/parameters/avatarStyleQuery'
        - $ref: '#/components/parameters/avatarSizeQuery'
      responses:
        '200':
          description: Avatar image bytes.
          headers:
            X-UOA-Avatar-Source:
              description: Which source the bytes came from (`provider` = the proxied team `iconUrl`).
              schema:
                type: string
                enum: [uploaded, provider, generated]
            Cache-Control:
              description: |
                `private, max-age=300` for uploaded/proxied images;
                `private, max-age=86400` for deterministic generated images.
              schema:
                type: string
            ETag:
              description: SHA-256 of the returned bytes.
              schema:
                type: string
            X-Content-Type-Options:
              schema:
                type: string
                enum: [nosniff]
            Content-Disposition:
              schema:
                type: string
            Content-Security-Policy:
              description: SVG responses only.
              schema:
                type: string
          content:
            image/svg+xml:
              schema:
                type: string
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '304':
          description: Not modified (matching `If-None-Match`).
        '401':
          description: Missing auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a superuser or wrong admin domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Avatars]
      summary: Set any team's uploaded avatar (admin view)
      operationId: adminPutTeamAvatar
      description: |
        Replaces the team's uploaded avatar so operators can manage company avatars from the
        Admin panel. Audit-logged as `team.avatar_updated` against the owning domain.
        Rate-limited per team.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AvatarUploadRequest'
      responses:
        '200':
          description: Stored avatar metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '400':
          description: Generic error — not a supported raster image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a superuser or wrong admin domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Upload exceeds the 1 MiB cap.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Avatars]
      summary: Remove any team's uploaded avatar (admin view)
      operationId: adminDeleteTeamAvatar
      description: |
        Idempotent. Resolution falls back to the team `iconUrl` or the generated image.
        Audit-logged as `team.avatar_deleted` against the owning domain. Rate-limited per team.
      security:
        - adminBearer: []
      responses:
        '200':
          description: Removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkTrue'
        '401':
          description: Missing auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a superuser or wrong admin domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...