# API/src/routes/org/groups.ts
import type { FastifyInstance, FastifyRequest } from 'fastify';
import { z } from 'zod';

import { asPrismaClient } from '../../db/tenant-context.js';
import { configVerifier } from '../../middleware/config-verifier.js';
import { requireOrgRole } from '../../middleware/org-role-guard.js';
import requireDomainHashAuthForDomainQuery from '../../middleware/domain-hash-auth.js';
import { requireOrgFeatures } from '../../middleware/org-features.js';
import { requireGroupsEnabled } from '../../middleware/groups-enabled.js';
import { setTenantContextFromRequest } from '../../plugins/tenant-context.plugin.js';
import { getGroup, listGroups } from '../../services/group.service.js';
import { AppError } from '../../utils/errors.js';
import { assertVerifiedDomainMatchesQuery, normalizeDomain } from './domain-context.js';

const DomainQuerySchema = z
  .object({
    domain: z
      .string()
      .trim()
      .min(1)
      .transform(normalizeDomain),
    config_url: z.string().trim().min(1),
  })
  .strict();

const ListQuerySchema = DomainQuerySchema.extend({
  limit: z.coerce.number().int().positive().max(200).optional(),
  cursor: z.string().trim().min(1).optional(),
}).strict();

const OrgPathSchema = z.object({
  orgId: z.string().trim().min(1),
});

const GroupPathSchema = OrgPathSchema.extend({
  groupId: z.string().trim().min(1),
});

function parseDomainContext(request: FastifyRequest) {
  const parsed = DomainQuerySchema.parse(request.query);
  assertVerifiedDomainMatchesQuery(request, parsed.domain);

  return parsed;
}

// preValidation runs before the handler's own strict query parse, so this hook
// must tolerate the list keys (limit, cursor) that the strict DomainQuerySchema
// would reject — brief §24.11 documents `?limit=&cursor=` on every list
// endpoint. Validate only the shared domain context here; each handler still
// strict-parses its full query schema.
const DomainContextHookSchema = DomainQuerySchema.passthrough();

async function parseDomainContextHook(request: FastifyRequest): Promise<void> {
  const parsed = DomainContextHookSchema.parse(request.query);
  assertVerifiedDomainMatchesQuery(request, parsed.domain);
}

function parseLimitCursor(request: FastifyRequest) {
  const parsed = ListQuerySchema.parse(request.query);
  assertVerifiedDomainMatchesQuery(request, parsed.domain);
  return parsed;
}

function getOrgIdFromParams(params: unknown): string {
  const parsed = OrgPathSchema.parse(params ?? {});
  return parsed.orgId;
}

function getGroupIdFromParams(params: unknown): string {
  const parsed = GroupPathSchema.parse(params ?? {});
  return parsed.groupId;
}

export function registerGroupRoutes(app: FastifyInstance): void {
  app.get(
    '/org/organisations/:orgId/groups',
    {
      preValidation: [
        requireDomainHashAuthForDomainQuery(),
        configVerifier,
        parseDomainContextHook,
        requireOrgFeatures,
        requireGroupsEnabled,
        requireOrgRole(),
      ],
    },
    async (request, reply) => {
      const { domain, limit, cursor } = parseLimitCursor(request);
      const config = request.config;
      if (!config) throw new AppError('UNAUTHORIZED', 401, 'MISSING_CONFIG');

      const orgId = getOrgIdFromParams(request.params);

      setTenantContextFromRequest(request, { orgId });
      const groups = await request.withTenantTx((tx) =>
        listGroups(
          { orgId, domain, config, limit, cursor },
          { prisma: asPrismaClient(tx) },
        ),
      );

      reply.status(200).send(groups);
    },
  );

  app.get(
    '/org/organisations/:orgId/groups/:groupId',
    {
      preValidation: [
        requireDomainHashAuthForDomainQuery(),
        configVerifier,
        parseDomainContextHook,
        requireOrgFeatures,
        requireGroupsEnabled,
        requireOrgRole(),
      ],
    },
    async (request, reply) => {
      const { domain } = parseDomainContext(request);
      const config = request.config;
      if (!config) throw new AppError('UNAUTHORIZED', 401, 'MISSING_CONFIG');

      const orgId = getOrgIdFromParams(request.params);
      const groupId = getGroupIdFromParams(request.params);

      setTenantContextFromRequest(request, { orgId });
      const group = await request.withTenantTx((tx) =>
        getGroup(
          { orgId, groupId, domain, config },
          { prisma: asPrismaClient(tx) },
        ),
      );

      reply.status(200).send(group);
    },
  );
}